refactor: select price range component
