+++ packages/web/src/components/common/select-price-range/select-price-range-custom/price-steps/PriceSteps.stories.tsx
@@ -1,0 +1,14 @@
+import SelectPriceRangeCutomController, { type PriceStepsProps } from "./PriceSteps";
+import { Meta, StoryObj } from "@storybook/react";
+
+export default {
+  title: "common/AddLiquidity/PriceSteps",
+  component: SelectPriceRangeCutomController,
+} as Meta<typeof SelectPriceRangeCutomController>;
+
+export const Default: StoryObj<PriceStepsProps> = {
+  args: {
+    current: 123123,
+    title: "GNOS per GNOT"
+  },
+};
